+++ packages/storybook-react/stories/extension-placeholder/basic.tsx
@@ -1,4 +1,4 @@
-import 'remirror/styles/all.css';
+import './styles.css';
 
 import React, { useCallback } from 'react';
 import { PlaceholderExtension } from 'remirror/extensions';
@@ -11,7 +11,30 @@ const Basic = (): JSX.Element => {
   );
   const { manager } = useRemirror({ extensions });
 
-  return <Remirror manager={manager} />;
+  return (
+    <ThemeProvider>
+      <Remirror manager={manager} />
+    </ThemeProvider>
+  );
+};
+
+export const CustomStyle = (): JSX.Element => {
+  const extensions = useCallback(
+    () => [
+      new PlaceholderExtension({
+        placeholder: `I'm a styled placeholder!`,
+        emptyNodeClass: 'my-placeholder',
+      }),
+    ],
+    [],
+  );
+  const { manager } = useRemirror({ extensions });
+
+  return (
+    <ThemeProvider>
+      <Remirror manager={manager} />
+    </ThemeProvider>
+  );
 };
 
 export default Basic;
