+++ packages/storybook-react/stories/extension-placeholder/placeholder-extension.stories.tsx
@@ -1,5 +1,4 @@
 import Basic from './basic';
-import WithCustomStyle from './with-custom-style';
 
-export { Basic, WithCustomStyle };
+export { Basic };
 export default { title: 'Extensions / Placeholder' };
