+++ packages/storybook-react/stories/extension-placeholder/styles.css
@@ -1,10 +1,3 @@
-.my-placeholder:first-of-type::before {
-  position: absolute;
-  pointer-events: none;
-  height: 0;
-  content: attr(data-placeholder);
-
-  color: lightcoral !important;
-  font-family: sans-serif;
-  font-weight: 900;
+.my-placeholder::before {
+  color: red !important;
 }
